requirements.txt
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-secret-manager==2.20.0
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
proto-plus==1.24.0
protobuf==4.25.3
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.32.3
rsa==4.9
Telethon==1.36.0
urllib3==2.2.2
